requirements.txt
python-jobspy>=1.1.82
fastapi>=0.115.0
uvicorn>=0.34.0
